+++ src/i18n.js
@@ -2,10 +2,14 @@ export const i18nValidator = {
   'en': {
     needAddr: 'Please pass the index/address or define SAIHUBOT_VALIDATOR first',
     query: 'Query validator\'s balance...',
+    accountBalance: 'Account Balance',
+    validatorBalance: '(Eth2) Validator Balance',
   },
   'zh_TW': {
     needAddr: '請傳入索引/地址，或是預先定義 SAIHUBOT_VALIDATOR 參數',
     query: '查詢驗證器餘額中...',
+    accountBalance: '地址餘額',
+    validatorBalance: '(Eth2) 驗證者餘額',
   },
   props: [],
 };